Create Check-SPOBrokenLink-link-linktext.ps1
This script helps SharePoint online intranet and digital workplace managers find broken links in their /sitepages/ aspx files.
It uses different methods to try and prove each type of link it finds in an ASPX page CanvasContent1 property. 
A CSV of the links, linktext, page the link is in and the script outcome is produced. This then needs to be manuall checked reviewed to get a final ist of broken links.